refactor: screen with barrel export
