+++ wabiz/client/wabiz/lib/views/favorite/favorite.dart
@@ -1,14 +1,1 @@
-import 'package:flutter/material.dart';
-
-class FavoriteScreen extends StatelessWidget {
-  const FavoriteScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('favorite'),
-      ),
-    );
-  }
-}
+export 'screen/favorite.dart';
