+++ wabiz/client/wabiz/lib/views/favorite/screen/favorite.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class FavoriteScreen extends StatelessWidget {
+  const FavoriteScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('favorite'),
+      ),
+    );
+  }
+}
