+++ wabiz/client/wabiz/lib/views/my/my.dart
@@ -1,14 +1,1 @@
-import 'package:flutter/material.dart';
-
-class MyScreen extends StatelessWidget {
-  const MyScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('my'),
-      ),
-    );
-  }
-}
+export 'screen/my.dart';
