+++ wabiz/client/wabiz/lib/views/my/screen/my.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class MyScreen extends StatelessWidget {
+  const MyScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('my'),
+      ),
+    );
+  }
+}
