+++ wabiz/client/wabiz/lib/views/project/project.dart
@@ -1,14 +1,1 @@
-import 'package:flutter/material.dart';
-
-class ProjectScreen extends StatelessWidget {
-  const ProjectScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('project'),
-      ),
-    );
-  }
-}
+export 'screen/project.dart';
