+++ wabiz/client/wabiz/lib/views/project/screen/project.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class ProjectScreen extends StatelessWidget {
+  const ProjectScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('project'),
+      ),
+    );
+  }
+}
